+++ CHANGELOG.md
@@ -64,6 +64,7 @@ All notable changes to this project will be documented in this file.
   to work on exclusive nodes.
 - Add better error reporting in setup script for invalid machine types.
 - Allow partial success in bulkInsert for resume.py
+- Add bulkInsert operation failure detection and logging
 
 ## \[4.1.5\]